UserOperation update.

+++ ScientificProgrammableCalculator/src/GUI/FXMLUserOperationsController.java
@@ -18,6 +18,7 @@ import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
+import javafx.scene.control.ButtonType;
 import javafx.scene.control.Label;
 import javafx.scene.control.MenuItem;
 import javafx.scene.control.TableColumn;
@@ -111,7 +112,7 @@ public class FXMLUserOperationsController implements Initializable {
        
         File file = fileChooser.showOpenDialog(null);
         if (file == null) {
-            //Implement alert
+            showAlert("Error", "Error in loading file", "You must specify a file.", Alert.AlertType.ERROR);
             return;
         }
         operations.clear();
@@ -140,7 +141,7 @@ public class FXMLUserOperationsController implements Initializable {
        
         File file = fileChooser.showSaveDialog(null);
         if (file == null) {
-            //Implement alert
+            showAlert("Error", "Error in saving file", "You must specify a file.", Alert.AlertType.ERROR);
             return;
         }
         
@@ -156,6 +157,7 @@ public class FXMLUserOperationsController implements Initializable {
         UserOperation selected = userOperationTable.getSelectionModel().getSelectedItem();
         if(selected == null)
             return;
+        
         TextInputDialog tid = new TextInputDialog();
         tid.setTitle("Edit");
         tid.setHeaderText("Edit selected user operation");
@@ -178,19 +180,22 @@ public class FXMLUserOperationsController implements Initializable {
         if(selected == null)
             return;
         
-        List<UserOperation> tmp = new ArrayList();
-        tmp.addAll(operations);
+        Alert deleteAlert = showAlert("Confirm removal", "", "All associated operations will be removed. Are you sure?", Alert.AlertType.CONFIRMATION);
+        if (deleteAlert.getResult() == ButtonType.OK) {
+            List<UserOperation> tmp = new ArrayList();
+            tmp.addAll(operations);
         
-        deleteOperation(selected, tmp);
+            deleteOperation(selected, tmp);
         
-        operations.clear();
-        operations.addAll(tmp);
-        operations.remove(selected);
-        
-        userOperationTable.setItems(operations);
-        mainReference.updateUserOperations(operations);
-        
-        calc.removeOperation(selected.getName());
+            operations.clear();
+            operations.addAll(tmp);
+            operations.remove(selected);
+
+            userOperationTable.setItems(operations);
+            mainReference.updateUserOperations(operations);
+
+            calc.removeOperation(selected.getName());
+        }
     }
     
     @FXML
@@ -212,12 +217,13 @@ public class FXMLUserOperationsController implements Initializable {
         userOperationTable.getItems().addAll(operations);
     }
     
-    private void showAlert(String title, String header, String content, Alert.AlertType type) {
+    private Alert showAlert(String title, String header, String content, Alert.AlertType type) {
         Alert alert = new Alert(type);
         alert.setTitle(title);
         alert.setContentText(content);
         alert.setHeaderText(header);
         alert.showAndWait();
+        return alert;
     }
     
     private void deleteOperation (UserOperation selected, List<UserOperation> tmp) {
